requirements.txt
absl-py==0.15.0
astunparse==1.6.3
bcrypt==4.0.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
cmake==3.27.0
contourpy==1.1.0
cryptography==41.0.2
cycler==0.11.0
filelock==3.12.2
Flask==2.3.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
flatbuffers==1.12
fonttools==4.41.1
gast==0.4.0
gevent==23.7.0
google-auth==2.22.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.34.1
gunicorn==20.1.0
h5py==3.1.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.7.0
keras-nightly==2.5.0.dev2021032800
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==16.0.6
lit==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
networkx==3.1
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-ml-py==12.535.77
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
nvitop==1.2.0
oauthlib==3.2.2
opencv-python==4.5.2.52
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
paramiko==3.2.0
Pillow==8.4.0
protobuf==3.20.3
psutil==5.9.5
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.6
pycparser==2.21
PyJWT==2.8.0
PyNaCl==1.5.0
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.11.1
seaborn==0.12.2
six==1.15.0
sshtunnel==0.4.0
sympy==1.12
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==1.1.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
triton==2.0.0
typing-extensions==3.7.4.3
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.3.6
wrapt==1.12.1
zipp==3.16.2
zope.event==5.0
zope.interface==6.0
